+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Add histograms comparing risk treatments
 
+* Update R Markdown template to match sample report
+
 # quantrr 1.2.1
 
 * Further reduce number of simulations in `calc_risk()` to resolve `code-link` issue after adding risk treatments